requirements.txt
requests>=2.31.0
streamlit>=1.28.0
pandas>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
jinja2>=3.1.0
faker>=20.0.0
pyyaml>=6.0
click>=8.1.0
rich>=13.0.0
langchain>=0.1.0
langchain-community>=0.0.30
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2 